requirements.txt
certifi==2019.6.16
olefile==0.46
Pillow==6.1.0
wincertstore==0.2